bugfix: no adding/editing/deleting foreign extensions/sipusers anymore

+++ voip/views.py
@@ -22,14 +22,14 @@ def pwgen(size=16):
 
 @login_required
 def index(request):
-    logger.debug('index accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('index accessed from %s by %s' % (request.META.get('REMOTE_ADDR'), request.user.username) )
     member = members.objects.get(nickname=request.user.username)
     context = {'request': request, 'member': member}
     return render(request, 'voip/index.html', context)
 
 @login_required
 def userinfo(request):
-    logger.debug('userinfo accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('userinfo accessed from %s by %s' % (request.META.get('REMOTE_ADDR'), request.user.username) )
     member = members.objects.get(nickname=request.user.username)
     try:
         ldapuserprofile = UserProfile.objects.get(uid=request.user.username)
@@ -40,7 +40,7 @@ def userinfo(request):
 
 @login_required
 def phonebook(request):
-    logger.debug('phonebook accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('phonebook accessed from %s by %s' % (request.META.get('REMOTE_ADDR'), request.user.username) )
     member = members.objects.get(nickname=request.user.username)
     extension_list = extensions.objects.filter(phonebook=True).order_by('extension')
     context = {'request': request, 'member': member, 'extension_list': extension_list}
@@ -48,7 +48,7 @@ def phonebook(request):
 
 @login_required
 def extensions_show(request):
-    logger.debug('extension_show accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('extension_show accessed from %s by %s' % (request.META.get('REMOTE_ADDR'), request.user.username) )
     member = members.objects.get(nickname=request.user.username)
     extension_list = extensions.objects.filter(id_members=member.id).order_by('extension')
     context = {'request': request, 'member': member, 'extension_list': extension_list,}
@@ -56,7 +56,7 @@ def extensions_show(request):
 
 @login_required
 def extension_add(request):
-    logger.debug('extension_add accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('extension_add accessed from %s by %s' % (request.META.get('REMOTE_ADDR'), request.user.username) )
     member = members.objects.get(nickname=request.user.username)
     if request.method == 'POST':
         form = ExtensionEditForm(request.POST)
@@ -77,26 +77,27 @@ def extension_add(request):
 
 @login_required
 def extension_edit(request, extension_id):
-    logger.debug('extension_edit accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('extension_edit(%s) accessed from %s by %s' % (extension_id, request.META.get('REMOTE_ADDR'), request.user.username) )
     member = members.objects.get(nickname=request.user.username)
     extension = extensions.objects.get(id=extension_id)
-    if request.method == 'POST':
-        form = ExtensionEditForm(request.POST)
-        if form.is_valid():
-            extension.voicemail = form.cleaned_data['voicemail']
-            extension.voicemail_pin = form.cleaned_data['voicemail_pin']
-            extension.voicemail_delay = form.cleaned_data['voicemail_delay']
-            extension.voicemail_emailnotify = form.cleaned_data['voicemail_emailnotify']
-            extension.phonebook = form.cleaned_data['phonebook']
-            extension.save()
-            return HttpResponseRedirect('/extensions/')
-    else:
-        form = ExtensionEditForm(initial={
-            'voicemail': extension.voicemail,
-            'voicemail_pin': extension.voicemail_pin,
-            'voicemail_delay': extension.voicemail_delay,
-            'voicemail_emailnotify': extension.voicemail_emailnotify,
-        })
+    if extension.id_members.nickname == member.nickname:
+        if request.method == 'POST':
+            form = ExtensionEditForm(request.POST)
+            if form.is_valid():
+                extension.voicemail = form.cleaned_data['voicemail']
+                extension.voicemail_pin = form.cleaned_data['voicemail_pin']
+                extension.voicemail_delay = form.cleaned_data['voicemail_delay']
+                extension.voicemail_emailnotify = form.cleaned_data['voicemail_emailnotify']
+                extension.phonebook = form.cleaned_data['phonebook']
+                extension.save()
+                return HttpResponseRedirect('/extensions/')
+        else:
+            form = ExtensionEditForm(initial={
+                'voicemail': extension.voicemail,
+                'voicemail_pin': extension.voicemail_pin,
+                'voicemail_delay': extension.voicemail_delay,
+                'voicemail_emailnotify': extension.voicemail_emailnotify,
+            })
     extension_list = extensions.objects.filter(id_members=member.id).order_by('extension')
     context = {'request': request, 'member': member, 'extension': extension, 'extension_list': extension_list, 'form': form}
     return render(request, 'voip/extension_edit.html', context)
@@ -104,47 +105,51 @@ def extension_edit(request, extension_id):
 @login_required
 def extension_delete(request, extension_id):
     extension = extensions.objects.get(id=extension_id)
-    logger.debug('extension_delete(%s) accessed from %s by %s' % (extension.extension,request.META.get('REMOTE_ADDR'),request.user.username) )
-    if not extension.astrt_sipusers_set.exists():
-        extension.id_members = None
-        extension.voicemail = False
-        extension.voicemail_pin = "1234"
-        extension.voicemail_delay = 15
-        extension.emailnotify = False
-        extension.phonebook = False
-        extension.save()
-        return HttpResponseRedirect('/extensions/')
-    else:
-      error = "Kann Durchwahl %s nicht entfernen, da noch %s SIP Devices vorhanden sind." % (extension.extension,extension.astrt_sipusers_set.count())
-      member = members.objects.get(nickname=request.user.username)
-      extension_list = extensions.objects.filter(id_members=member.id).order_by('extension')
-      context = {'request': request, 'member': member, 'extension_list': extension_list, 'error': error,}
-      return render(request, 'voip/extensions.html', context)
+    logger.debug('extension_delete(%s) accessed from %s by %s' % (extension.extension, request.META.get('REMOTE_ADDR'), request.user.username) )
+    member = members.objects.get(nickname=request.user.username)
+    if extension.id_members.nickname == member.nickname:
+        if not extension.astrt_sipusers_set.exists():
+            extension.id_members = None
+            extension.voicemail = False
+            extension.voicemail_pin = "1234"
+            extension.voicemail_delay = 15
+            extension.emailnotify = False
+            extension.phonebook = False
+            extension.save()
+            return HttpResponseRedirect('/extensions/')
+        else:
+            error = "Kann Durchwahl %s nicht entfernen, da noch %s SIP Devices vorhanden sind." % (extension.extension, extension.astrt_sipusers_set.count() )
+            member = members.objects.get(nickname=request.user.username)
+            extension_list = extensions.objects.filter(id_members=member.id).order_by('extension')
+            context = {'request': request, 'member': member, 'extension_list': extension_list, 'error': error,}
+            return render(request, 'voip/extensions.html', context)
 
 @login_required
 @transaction.commit_manually
 def sipuser_add(request, extension_id):
-    logger.debug('sipuser_add accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('sipuser_add on extension %s accessed from %s by %s' % (extension_id, request.META.get('REMOTE_ADDR'), request.user.username) )
+    member = members.objects.get(nickname=request.user.username)
     extension = extensions.objects.get(id=extension_id)
     nickname = request.user.username
-    from django.db import connection
-    cursor = connection.cursor()
-    secret = pwgen()
-    cursor.execute("SELECT func_create_sipuser('%s','%s',%s)" % (nickname,secret,extension_id) )
-    ret = cursor.fetchone()
-    logger.debug("extension_id: %s - stored procedure 'func_create_sipuser' returned: %s" % (extension_id,ret) )
-    pin = 1234
-    timeout = 15
-    cursor.execute("SELECT func_create_voicemail_from_phonenumberid(%s,'%s',%s)" % (extension_id, pin, timeout) )
-    ret = cursor.fetchone()
-    logger.debug("stored procedure 'func_create_voicemail' returned: %s" % ret)
-    cursor.close() # is this needed and/or on the right position called?
-    connection.commit()
+    if extension.id_members.nickname == member.nickname:
+        from django.db import connection
+        cursor = connection.cursor()
+        secret = pwgen()
+        cursor.execute("SELECT func_create_sipuser('%s','%s',%s)" % (nickname, secret, extension_id) )
+        ret = cursor.fetchone()
+        logger.debug("extension_id: %s - stored procedure 'func_create_sipuser' returned: %s" % (extension_id, ret) )
+        pin = 1234
+        timeout = 15
+        cursor.execute("SELECT func_create_voicemail_from_phonenumberid(%s,'%s',%s)" % (extension_id, pin, timeout) )
+        ret = cursor.fetchone()
+        logger.debug("stored procedure 'func_create_voicemail' returned: %s" % ret)
+        cursor.close() # is this needed and/or on the right position called?
+        connection.commit()
     return HttpResponseRedirect('/extensions/')
 
 @login_required
 def sipuser_edit(request, sipuser_id):
-    logger.debug('sipuser_edit accessed from %s by %s' % (request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('sipuser_edit(%s) accessed from %s by %s' % (sipuser_id, request.META.get('REMOTE_ADDR'), request.user.username) )
     member = members.objects.get(nickname=request.user.username)
     sipuser = astrt_sipusers.objects.get(id=sipuser_id)
     if request.method == 'POST':
@@ -174,7 +179,9 @@ def sipuser_edit(request, sipuser_id):
 
 @login_required
 def sipuser_delete(request, sipuser_id):
-    logger.debug('sipuser_delete(%s) accessed from %s by %s' % (sipuser_id,request.META.get('REMOTE_ADDR'),request.user.username) )
+    logger.debug('sipuser_delete(%s) accessed from %s by %s' % (sipuser_id, request.META.get('REMOTE_ADDR'), request.user.username) )
+    member = members.objects.get(nickname=request.user.username)
     sipuser = astrt_sipusers.objects.get(id=sipuser_id)
-    sipuser.delete()
+    if sipuser.phonenumberid.id_members.nickname == member.nickname:
+        sipuser.delete()
     return HttpResponseRedirect('/extensions/')
